+++ examples/weather/shared/src/effects/http/mod.rs
@@ -1,0 +1,2 @@
+pub mod location;
+pub mod weather;
