+++ examples/weather/shared/src/effects/http/weather/model/current_response.rs
