+++ examples/weather/shared/src/effects/http/weather/model/mod.rs
