+++ examples/weather/shared/src/effects/http/weather/model/response_elements/clouds.rs
