+++ examples/weather/shared/src/effects/http/weather/model/response_elements/mod.rs
