+++ examples/weather/shared/src/effects/mod.rs
@@ -1,3 +1,4 @@
+pub mod http;
 pub mod location;
 pub mod secret;
