--- examples/weather/shared/src/model/active/home/weather/mod.rs
@@ -1,2 +1,0 @@
-pub mod client;
-pub mod model;
